Updates for recent nuc install
Made prompt used for pexpect less fragile, as something ANSI codes in
the prompt were changing (after setting locale perhaps) that made it
not detect it.

pip install the required Python libraries - using Pacman is no good due
to a partial upgrade situation when Python has recently upgraded in the
arch repos.

Change to the number corresponding to a partition type in fdisk.